+++ listings/ch11/listing_07.js
@@ -1,36 +1,7 @@
-export function patchDOM(oldVdom, newVdom, parentEl, hostComponent = null) {
-  if (!areNodesEqual(oldVdom, newVdom)) {
-    const index = findIndexInParent(parentEl, oldVdom.el)
-    destroyDOM(oldVdom)
-    mountDOM(newVdom, parentEl, index, hostComponent)
+function patchComponent(oldVdom, newVdom) {
+  const { component } = oldVdom
+  const { props: newProps } = newVdom
 
-    return newVdom
-  }
-
-  newVdom.el = oldVdom.el
-
-  switch (newVdom.type) {
-    case DOM_TYPES.TEXT: {
-      patchText(oldVdom, newVdom)
-      return newVdom
-    }
-
-    case DOM_TYPES.ELEMENT: {
-      patchElement(oldVdom, newVdom, hostComponent)
-      break
-    }
-
-    // --add--
-    case DOM_TYPES.COMPONENT: {
-      patchComponent(oldVdom, newVdom)
-      break
-    }
-    // --add--
-  }
-
-  patchChildren(oldVdom, newVdom, hostComponent)
-
-  return newVdom
+  newVdom.component = component
+  component.updateProps(newProps)
 }
-
-// TODO: implement patchComponent()
