+++ listings/ch11/listing_11.js
@@ -1,15 +1,12 @@
-function createComponentNode(vdom, parentEl, index, hostComponent) {
-  const Component = vdom.tag
-  // --remove--
-  const props = extractComponentProps(vdom)
-  const component = new Component(props)
-  // --remove--
-  // --add--
-  const { props, events } = extractComponentProps(vdom)
-  const component = new Component(props, events, hostComponent)
-  // --add--
+export function extractComponentProps(vdom) {
+  const { on: events = {}, ...props } = vdom.props
+  delete props.key
 
-  component.mount(parentEl, index)
-  vdom.component = component
-  vdom.el = component.firstElement
+  for (const prop in props) {
+    if (prop.startsWith('data-')) {
+      delete props[prop]
+    }
+  }
+
+  return { props, events }
 }
